requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
joblib==1.4.2
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
Werkzeug==3.0.3